requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
openai==1.3.0
Flask-Session==0.5.0
requests>=2.32.3
